refactored out common text field
